requirements.txt
torch
transformers
sentence-transformers
streamlit
PyPDF2
scikit-learn
numpy
faiss-cpu
